==> 03.retrieve-annotations/0.0.phenotypic-activity-orf-CRISPRPipeline.ipynb <==
import pandas as pd
import utils
import numpy as np
import warnings
from tqdm import tqdm
from copairs.map import average_precision, mean_average_precision
import logging

logging.basicConfig(format="%(levelname)s:%(asctime)s:%(name)s:%(message)s")
logging.getLogger("copairs").setLevel(logging.INFO)

warnings.simplefilter(action="ignore", category=FutureWarning)

#operations = "wellpos_cc_var_mad_outlier_featselect_sphering_harmony"
batch_size = 20000
null_size = 20000
fdr = 0.1

### Prepare the data

#### Read the CRISPR parquet file

#Reading the output from the ORF pipeline
raw_CRISPR_df = pd.read_parquet('C:\\Users\\ssivagur\\Documents\\GitHub\\ssivagur\\FeatureSpaceIntegration\\\CRISPRPipelineValidation\\profiles_wellpos_cc_var_mad_outlier_featselect_sphering_harmony_PCA_corrected.parquet')

raw_CRISPR_df.shape

# Filtering only the ORF from the raw output 
orf_df = raw_CRISPR_df[raw_CRISPR_df['Metadata_PlateType'] == 'ORF']

orf_df.shape

#dropping the duplicate metadata columns
orf_df_cols = [c for c in orf_df.columns if c.startswith("Metadata_")]

orf_df_cols

# there are duplicate columns of NCBI_Gene_ID and the Symbol and the columns with the NaN are dropped
# orf_df['Metadata_NCBI_Gene_ID_y'] has NaN and it is dropped
orf_df['Metadata_NCBI_Gene_ID_y']

orf_df['Metadata_NCBI_Gene_ID_y'].isnull().all()

orf_df = orf_df.drop(['Metadata_NCBI_Gene_ID_y'], axis=1)

# orf_df['Metadata_Symbol_y'] has NaN and it is dropped
orf_df['Metadata_Symbol_y'].isnull().all()

# orf_df['Metadata_Symbol'] has NaN and it is dropped
orf_df['Metadata_Symbol'].isnull().all()

orf_df = orf_df.drop(['Metadata_Symbol_y', 'Metadata_Symbol'], axis=1)

orf_df.shape

#Renaming the ['Metadata_Symbol_y'] and ['Metadata_NCBI_Gene_ID_y'] to be used in the further steps 
orf_df = orf_df.rename(columns={'Metadata_NCBI_Gene_ID_x':'Metadata_NCBI_Gene_ID', 'Metadata_Symbol_x':'Metadata_Symbol'})

orf_df.shape

orf_df.to_parquet('C:\\Users\\ssivagur\\Documents\\GitHub\\ssivagur\\FeatureSpaceIntegration\\CRISPRPipelineValidation\\profiles_ORF.parquet')

#### Add annotations - I did not run this since the profiles that I have already has the annotations in it

#orf_metdata_df = pd.read_csv(
    #"../00.download-and-process-annotations/output/orf_metadata.tsv.gz", sep="\t"
#)
#compound_metadata_df = pd.read_csv(
    #"../datasets/metadata/compound.csv.gz", usecols=["Metadata_JCP2022"]
#).#assign(
    #Metadata_pert_type=lambda x: np.where(
     #   x["Metadata_JCP2022"] == "JCP2022_999999", "empty", "poscon"
 #   )
#)

#metadata_df = pd.concat(
 #   [
        #orf_metdata_df,
        #compound_metadata_df
  #  ],
    #join="outer",
    #ignore_index=True,
#)

#orf_df = orf_df.merge(metadata_df, on="Metadata_JCP2022", how="inner")
#orf_df.shape

#### Remove empty wells

orf_df = utils.remove_empty_wells(orf_df)
orf_df.shape

#### Remove `poscon` wells.

orf_df = orf_df.query('Metadata_pert_type!="poscon"').reset_index(drop=True)
orf_df.shape

#### Remove `BAD CONSTRUCT` profiles

orf_df = orf_df.query('Metadata_broad_sample!="BAD CONSTRUCT"').reset_index(drop=True)
orf_df.shape

#### Remove featues with `nan` values.
These need to be removed as the `nan` values will cause the mean average precision calculation to fail.

orf_df = utils.remove_nan_features(orf_df)

#### Remove low infection effiency wells

# Add platemap name

platemap_df = pd.read_csv(
    "../00.download-and-process-annotations/input/experiment-metadata.tsv",
    sep="\t",
    usecols=["Plate_Map_Name", "Assay_Plate_Barcode"],
).rename(
    columns={
        "Plate_Map_Name": "Metadata_plate_map_name",
        "Assay_Plate_Barcode": "Metadata_Plate",
    }
)

orf_df = orf_df.merge(platemap_df, on="Metadata_Plate", how="left")

orf_df = utils.remove_low_infection_efficiency_wells(orf_df)
orf_df.shape

### Calculate mAP for each ORF perturbation

# Adding a new column for negative control
orf_df["Metadata_negcon"] = np.where(orf_df["Metadata_pert_type"] == "negcon", 1, 0)

pos_sameby = ["Metadata_JCP2022"]
pos_diffby = []
neg_sameby = ["Metadata_Plate"]
neg_diffby = ["Metadata_negcon"]

metadata_df = utils.get_metadata(orf_df)
feature_df = utils.get_featuredata(orf_df)
feature_values = feature_df.values

result = average_precision(
    metadata_df, feature_values, pos_sameby, pos_diffby, neg_sameby, neg_diffby, batch_size=batch_size
)

# Remove negcon
result = result.query('Metadata_pert_type!="negcon"').reset_index(drop=True)

agg_result = (
    mean_average_precision(result, pos_sameby, null_size=null_size, threshold=fdr, seed=12527)
    .rename(columns={'average_precision': 'mean_average_precision'})
)

agg_result.to_csv(f"C:\\Users\\ssivagur\\Documents\\GitHub\\ssivagur\\FeatureSpaceIntegration\\CRISPRPipelineValidation\\PhenotypicActivity_ORF.csv.gz", index=False)
