==> 03.retrieve-annotations/1.0.phenotypic-activity-crispr-CRISPRPipeline.ipynb <==
import pandas as pd
import utils
import numpy as np
import warnings
from tqdm import tqdm
from copairs.map import average_precision, mean_average_precision
import logging

logging.basicConfig(format="%(levelname)s:%(asctime)s:%(name)s:%(message)s")
logging.getLogger("copairs").setLevel(logging.INFO)

warnings.simplefilter(action="ignore", category=FutureWarning)

#operations = "wellpos_cc_var_mad_outlier_featselect_sphering_harmony_PCA_corrected"
batch_size = 20000
null_size = 20000
fdr = 0.1

# CRISPR controls - did not use this since the profies already has pert_type in it

#crispr_controls_df = pd.DataFrame(
#    {
#        "Metadata_JCP2022": ["JCP2022_805264", "JCP2022_800001", "JCP2022_800002"],
#        "Metadata_pert_type": ["poscon", "negcon", "negcon"],
##
#    index=[0, 1, 2],
#)

### Prepare the data

#### Filtering the CRISPR plates from the output file

raw_CRISPR_df = pd.read_parquet('C:\\Users\\ssivagur\\Documents\\GitHub\\ssivagur\\FeatureSpaceIntegration\\CRISPRPipelineValidation\\profiles_wellpos_cc_var_mad_outlier_featselect_sphering_harmony_PCA_corrected.parquet')

raw_CRISPR_df.shape

#filtering 
crispr_df = raw_CRISPR_df[raw_CRISPR_df['Metadata_PlateType'] == 'CRISPR']

There are two columns of metadata for 'Metadata_Symbol' in the raw_orf_df - 'Metadata_Symbol_x' and 'Metadata_Symbol_y'. I have to check why there are two columns 

crispr_df.shape

crispr_df_cols = [c for c in crispr_df.columns if c.startswith("Metadata_")]

crispr_df_cols

crispr_df['Metadata_NCBI_Gene_ID_x']

crispr_df['Metadata_NCBI_Gene_ID_x'].isnull().all()

crispr_df['Metadata_Symbol'].isnull().all()

crispr_df['Metadata_Symbol_x'].isnull().all()


crispr_df['Metadata_Symbol_y'].isnull().all()


len(crispr_df['Metadata_Symbol_y'].notna())

len(crispr_df['Metadata_Symbol'].notna())

non_na_with_locs = crispr_df.where(crispr_df['Metadata_Symbol'].notna())

non_na_with_locs

crispr_df['Metadata_Symbol_y'] and crispr_df['Metadata_Symbol'] is having around 5636 NaN which corresponds to 8 compound poscons and DMSO which can be filtered. Following are the JCP ID's of the 8 compound poscons and DMSO
['JCP2022_033924',
 'JCP2022_050797',
 'JCP2022_025848',
 'JCP2022_046054',
 'JCP2022_037716',
 'JCP2022_064022',
 'JCP2022_012818',
 'JCP2022_085227',
 'JCP2022_035095']

#Filtering/dropping the JCP ID's corresponding to 8 compound poscons and DMSO
crispr_df = crispr_df[~(crispr_df["Metadata_JCP2022"].isin(['JCP2022_033924',
 'JCP2022_050797',
 'JCP2022_025848',
 'JCP2022_046054',
 'JCP2022_037716',
 'JCP2022_064022',
 'JCP2022_012818',
 'JCP2022_085227',
 'JCP2022_035095']))]

len(crispr_df[(crispr_df["Metadata_JCP2022"].isin(['JCP2022_033924',
 'JCP2022_050797',
 'JCP2022_025848',
 'JCP2022_046054',
 'JCP2022_037716',
 'JCP2022_064022',
 'JCP2022_012818',
 'JCP2022_085227',
 'JCP2022_035095']))])

'JCP2022_033924' in crispr_df['Metadata_Symbol_y'].values

crispr_df.shape

differences = crispr_df[crispr_df['Metadata_Symbol_y'] != crispr_df['Metadata_Symbol']]
print(differences)

only_two = differences[['Metadata_Symbol_y', 'Metadata_Symbol']]

only_two.isnull().all()

#dropping the columns ['Metadata_Symbol_x'] and ['Metadata_NCBI_Gene_ID_x'] since they have NaNs
crispr_df = crispr_df.drop(['Metadata_Symbol_x', 'Metadata_NCBI_Gene_ID_x', 'Metadata_Symbol'], axis=1)

#renaming the ['Metadata_Symbol_y'] and ['Metadata_NCBI_Gene_ID_y'] to be used downstream 
crispr_df = crispr_df.rename(columns={'Metadata_NCBI_Gene_ID_y':'Metadata_NCBI_Gene_ID', 'Metadata_Symbol_y':'Metadata_Symbol'})

crispr_df.shape

crispr_df.to_parquet('C:\\Users\\ssivagur\\Documents\\GitHub\\ssivagur\\FeatureSpaceIntegration\\CRISPRPipelineValidation\\profiles_crispr.parquet')

The size of the original file is (51185, 263)

#### Add annotations - not running this since the profiles alreayd has the metadata in it 

#    pd.read_csv(
#        "../00.download-and-process-annotations/output/crispr_metadata.tsv.gz", sep="\t"
#    )
#   .merge(crispr_controls_df, on="Metadata_JCP2022", how="left")
#    .fillna(value={"Metadata_pert_type": "trt"})
#)
#compound_metadata_df = pd.read_csv(
#    "../datasets/metadata/compound.csv.gz", usecols=["Metadata_JCP2022"]
#).assign(
#    Metadata_pert_type=lambda x: np.where(
#        x["Metadata_JCP2022"] == "JCP2022_999999", "empty", "poscon"
#    )
#)

#metadata_df = pd.concat(
#    [crispr_metdata_df, compound_metadata_df],
#    join="outer",
#    ignore_index=True,
#)

#crispr_df = crispr_df.merge(metadata_df, on="Metadata_JCP2022", how="inner")
#crispr_df.shape

#### Remove `poscon` wells.

crispr_df = crispr_df.query('Metadata_pert_type!="poscon"').reset_index(drop=True)
crispr_df.shape

#### Remove featues with `nan` values.
These need to be removed as the `nan` values will cause the mean average precision calculation to fail.

crispr_df = utils.remove_nan_features(crispr_df)
crispr_df.shape

#### Remove reagents with only one replicate

crispr_df.Metadata_JCP2022.value_counts()[
        crispr_df.Metadata_JCP2022.value_counts() == 1
    ].reset_index()['Metadata_JCP2022'].tolist()

#changed the code to the column name
reagents_with_one_replicate = (
    crispr_df.Metadata_JCP2022.value_counts()[
        crispr_df.Metadata_JCP2022.value_counts() == 1
    ]
    .reset_index()["Metadata_JCP2022"] #.reset_index()["index"]
    .tolist()
)



crispr_df = crispr_df[~crispr_df.Metadata_JCP2022.isin(reagents_with_one_replicate)]
crispr_df.shape

### Calculate mAP for each ORF perturbation

# Adding a new column for negative control
crispr_df["Metadata_negcon"] = np.where(crispr_df["Metadata_pert_type"] == "negcon", 1, 0)

pos_sameby = ["Metadata_JCP2022"]
pos_diffby = []
neg_sameby = ["Metadata_Plate"]
neg_diffby = ["Metadata_negcon"]

metadata_df = utils.get_metadata(crispr_df)
feature_df = utils.get_featuredata(crispr_df)
feature_values = feature_df.values

result = average_precision(
    metadata_df, feature_values, pos_sameby, pos_diffby, neg_sameby, neg_diffby, batch_size=batch_size
)

# Remove negcon
result = result.query('Metadata_pert_type!="negcon"').reset_index(drop=True)

agg_result = (
    mean_average_precision(result, pos_sameby, null_size=null_size, threshold=fdr, seed=12527)
    .rename(columns={'average_precision': 'mean_average_precision'})
)

agg_result.to_csv(f"C:\\Users\\ssivagur\\Documents\\GitHub\\ssivagur\\FeatureSpaceIntegration\\CRISPRPipelineValidation\\PhenotypicActivity_CRISPR.csv.gz", index=False)
